Make discovery server optional via Option<u16> port (#17)
Change Server.discovery_port from u16 to Option<u16>. When set to None,
the discovery server is not started, which is useful for tests and
environments where ASCOM discovery is not needed.

This also fixes the dead field bug where Server.discovery_port was never
actually used — Server::bind() always hardcoded DEFAULT_DISCOVERY_PORT
(32227) via DiscoveryServer::for_alpaca_server_at(). Now the configured
port is passed through to the discovery server.

Breaking change: discovery_port is now Option<u16> (default Some(32227)).
discovery_listen_addr() now returns Option<SocketAddr>.


* Make DEFAULT_DISCOVERY_PORT public

Expose the ASCOM Alpaca default discovery port (32227) as a public
constant so downstream crates can reference it instead of duplicating
the magic number.

+++ examples/camera-server/src/main.rs
@@ -587,7 +587,7 @@ fn get_webcam(camera_info: &CameraInfo) -> eyre::Result<Webcam> {
             let bin_x = exact_div(max_format_res.x, format_res.x)?;
             let bin_y = exact_div(max_format_res.y, format_res.y)?;
 
-            (bin_x == bin_y).then_some(bin_x as u8)
+            (bin_x == bin_y).then_some(bin_x)
         })
         .collect::<Vec<_>>();
